refactor: extract 51job search utility functions into lib module (Phase 3a)
Extract 15 pure utility functions from content.ts into
lib/job51-search-extractor.ts using DI factory pattern, matching the
seek/job5156 extraction approach. Removed SyncStatusWidget dependency
from waitForJob51Cooldown to avoid TDZ initialization issue (the widget
call was cosmetic UX only; cooldown logic preserved).

Functions extracted: isJob51DetailPage, isJob51DetailReady,
normalizeJob51AuthContext, getJob51RawRows, getJob51TotalFromPayload,
isLikelyJob51ResumeRow, getJob51ResumeRows, hasJob51SearchSnapshot,
isJob51EmptySearchPromptVisible, isJob51RateLimitedPage,
ensureJob51PageAllowed, waitForJob51Cooldown,
isJob51RateLimitedErrorMessage, isJob51RateLimitedPayload,
isJob51DetailApiErrorPayload.

content.ts: 6079 → 5893 lines (-186)
new file: lib/job51-search-extractor.ts (301 lines)
